+++ GraceNotes/GraceNotes/Application/GraceNotesApp.swift
@@ -34,6 +34,7 @@ struct GraceNotesApp: App {
         if !isRunningUnitTests {
             JournalTutorialStorageKeys.migrateLegacyKeysIfNeeded(using: .standard)
             JournalAppearanceMode.migrateLegacySummerRawValueIfNeeded(defaults: .standard)
+            AccentPreference.migrateRemovedCasesIfNeeded(using: .standard)
             _ = ICloudSyncPreferenceResolver.resolvedCloudSyncEnabled(using: .standard)
             JournalOnboardingProgress.migrateLegacyPostSeedOrientationFlagsIfNeeded(using: .standard)
             JournalOnboardingProgress.migrateLegacyAppTourSeenFlagIfNeeded(using: .standard)